I abstract functions from React components to lib / util files to follow single-responsibility-principle.
