+++ .env.example
@@ -11,3 +11,4 @@ ALCHEMY_API_KEY=
 NEXT_PUBLIC_NFT_STORAGE_TOKEN=
 NEXT_PUBLIC_MUSIC_GAME_CONTRACT_ADDRESS=
 ABLY_API_KEY=
+NEXT_PUBLIC_RPC_URL=
