+++ lib/getProvider.ts
@@ -1,0 +1,8 @@
+import { ethers } from "ethers"
+
+const getProvider = () => {
+  const provider = ethers.getDefaultProvider(process.env.NEXT_PUBLIC_RPC_URL)
+  return provider
+}
+
+export default getProvider
